+++ truffle/CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `TriState` utility class represents three states TRUE, FALSE and UNDEFINED.
 * Added `InteropLibrary.getUncached()` and `InteropLibrary.getUncached(Object)` short-cut methods for convenience.
 
+* Added `ConditionProfile#create()` as an alias of `createBinaryProfile()` so it can be used like `@Cached ConditionProfile myProfile`. 
 
 ## Version 20.1.0
 * Added `@GenerateLibrary(dynamicDispatchEnabled = false)` that allows to disable dynamic dispatch semantics for a library. The default is `true`.